experiments/vector_tag_documents/requirements.txt
# Core dependencies for BERTopic document vectorization experiment

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# PDF processing
PyPDF2>=3.0.0

# BERTopic and NLP
bertopic>=0.15.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Google Cloud Storage
google-cloud-storage>=2.10.0

# MLflow
mlflow>=2.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment variables
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
